Fix [ 1085454 ] Py 2.4 + MS Excel COM --> crash
Avoid crashes on Python 2.4 by defaulting to bForDemand=True - ie, each
typelib actually creates a "skeleton" Python package, and populates
the package as each interface is requested.

Cause of the underling crash is that the generated .py causes an overflow
as the byte-code is generated - something in 2.4 bloated the byte-code
for these modules.