Check for archive before assuming the module is readonly
Previously only zipimport modules had a __loader__ attribute at all
(and these always had an archive attribute). Now, all modules have
__loader__ items, representing the loader which loaded them. Only
some have archive attributes. There doesn't seem to be any canonical
mechanism for identifying a module as readonly, so we assume that
a loader with an archive attribute is.